\name{HMMsimulate}
\alias{HMMsimulate}
\title{HMMsimulate}
\usage{
HMMsimulate(tuningparameters)
}
\arguments{
  \item{tuningparameters}{Carried values.}
}
\value{
Auxiliary function called from C++.
}
\description{
Auxiliary function called from C++.
}

\references{
Yang Chen, Cheng-Der Fuh, Chu-Lan Kao, and Samuel Kou (2019+) "Determine the number of states
in hidden markov models via marginal likelihood." Submitted.
}

